--- idea/frontend/src/entities/waitlist/index.ts
@@ -1,5 +1,0 @@
-import { Interval, WaitlistContent, HumanWaitlistItem } from './model/types';
-import { HorizontalWaitlistItem } from './ui/horizontalWaitlistItem';
-
-export { HorizontalWaitlistItem };
-export type { Interval, WaitlistContent, HumanWaitlistItem };
