--- idea/frontend/src/entities/waitlist/ui/horizontalWaitlistItem/HorizontalWaitlistItem.module.scss
@@ -1,56 +1,0 @@
-@use '@/shared/assets/styles/_shared.scss' as *;
-@use '@/shared/assets/styles/_mixins.scss' as *;
-@use '@/shared/assets/styles/_variables.scss' as *;
-
-%textStyles {
-  color: $gray600;
-  font-size: $fontSizeTiny;
-}
-
-.horizontalMessageCard {
-  display: grid;
-  grid-template-rows: minmax(max-content, 1fr);
-  grid-template-columns: minmax(max-content, 1fr) minmax(max-content, 1.5fr) minmax(max-content, 0.5fr);
-  gap: 30px;
-  align-items: center;
-  position: relative;
-  padding: toRem(20) toRem(24);
-  background: $bgColorSecondary;
-  box-shadow: 0px 4px 4px $bgColor;
-  border-radius: 16px;
-
-  .entry {
-    @extend %textStyles;
-
-    justify-self: center;
-    line-height: 1.3;
-  }
-
-  .interval {
-    @include childrenMargin($marginSmall, right);
-    @extend %textStyles;
-
-    display: flex;
-    align-items: center;
-    font-weight: 500;
-    line-height: toRem(18);
-
-    .tooltip {
-      z-index: 2;
-
-      &::after {
-        left: 100%;
-        transform: translateX(-100%);
-      }
-    }
-  }
-
-  .link {
-    position: absolute;
-    border: inherit;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-  }
-}
