--- idea/frontend/src/entities/waitlist/ui/horizontalWaitlistItem/index.ts
@@ -1,3 +1,0 @@
-import { HorizontalWaitlistItem } from './HorizontalWaitlistItem';
-
-export { HorizontalWaitlistItem };
