+++ idea/frontend/src/features/message/index.ts
@@ -1,7 +1,7 @@
 import { getDecodedMessagePayload } from './utils';
-import { MessageCard } from './ui';
+import { ProgramMessages } from './ui';
 import { MESSAGE_TYPE } from './consts';
 import { Message } from './types';
 
-export { getDecodedMessagePayload, MessageCard, MESSAGE_TYPE };
+export { getDecodedMessagePayload, ProgramMessages, MESSAGE_TYPE };
 export type { Message };
