+++ idea/frontend/src/features/message/ui/program-messages/program-messages.module.scss
@@ -1,7 +1,7 @@
-@use '@/shared/assets/styles/_mixins.scss' as *;
-@use '@/shared/assets/styles/_shared.scss' as *;
-@use '@/shared/assets/styles/_variables.scss' as *;
-@use '@/shared/assets/styles/_animations.scss' as *;
+@use '@/shared/assets/styles/mixins' as *;
+@use '@/shared/assets/styles/shared' as *;
+@use '@/shared/assets/styles/variables' as *;
+@use '@/shared/assets/styles/animations' as *;
 
 .programMessages {
   position: sticky;
