--- idea/frontend/src/shared/ui/switchButton/SwitchButton.module.scss
@@ -1,23 +1,0 @@
-@use '@/shared/assets/styles/_shared.scss' as *;
-@use '@/shared/assets/styles/_variables.scss' as *;
-
-.switchButton {
-  color: $gray300;
-  padding: toRem(6) toRem(12);
-  font-size: $fontSizeSmall;
-  font-weight: 700;
-  font-family: 'Kanit';
-  line-height: toRem(14);
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-  background-color: transparent;
-
-  &:hover {
-    color: $successColor;
-  }
-
-  &.active {
-    color: $successColor;
-    background-color: $green007;
-  }
-}
