--- idea/frontend/src/shared/ui/switchButton/index.ts
@@ -1,3 +1,0 @@
-import { SwitchButton } from './SwitchButton';
-
-export { SwitchButton };
