--- idea/frontend/src/widgets/programMessages/index.ts
@@ -1,3 +1,0 @@
-import { ProgramMessages } from './ui/ProgramMessages';
-
-export { ProgramMessages };
