--- idea/frontend/src/widgets/programMessages/model/consts.ts
@@ -1,6 +1,0 @@
-enum MessageFilter {
-  Messages = 'messages',
-  Waitlist = 'waitlist',
-}
-
-export { MessageFilter };
