--- idea/frontend/src/widgets/programMessages/ui/messages/index.ts
@@ -1,3 +1,0 @@
-import { Messages } from './Messages';
-
-export { Messages };
