--- idea/frontend/src/widgets/programMessages/ui/waitlist/index.ts
@@ -1,3 +1,0 @@
-import { Waitlist } from './Waitlist';
-
-export { Waitlist };
